:minibus: Move tree-view into it's own dir
