+++ lib/tree-view/index.js
@@ -5,7 +5,7 @@
 import Path from 'path'
 import { CompositeDisposable, Emitter } from 'sb-event-kit'
 import { projectPathByFile } from './helpers'
-import type { Message, MessageLegacy, MessagesPatch, TreeViewHighlights, TreeViewPatch } from './types'
+import type { Message, MessageLegacy, MessagesPatch, TreeViewHighlights, TreeViewPatch } from '../types'
 
 export default class TreeView {
   emitter: Emitter;
@@ -29,7 +29,7 @@ export default class TreeView {
         this.decorateOnTreeView = decorateOnTreeView
       } else {
         const messages = this.messages
-        this.apply({ added: [], messages: [], removed: this.messages })
+        this.apply({ added: [], messages: [], removed: messages })
         this.decorateOnTreeView = decorateOnTreeView
         this.apply({ added: messages, messages, removed: [] })
       }
